+++ CHANGES.md
@@ -8,6 +8,7 @@
 - #908 Login form for inVerse is only 200px when `allow_registration` is set to `false`.
 - #909 Translations written as template literals [aren't parsed properly by xgettext](https://savannah.gnu.org/bugs/?50920).
 - #911 Use `getDefaultNickName` consistently to allow better overrides via plugins.
+- #912 `maximize` method in `converse-minimize` fails if the `controlbox` is not there.
 
 ## 3.2.0 (2017-08-09)